1. Change id from String or Float to Interger 2. Relationship @ManyToMany for Person and Role 3. Remove duplicate bar code in Course 4. Change default password to root

+++ src/main/java/edu/cs544/team5/domain/Location.java
@@ -11,7 +11,8 @@ import javax.persistence.*;
 public class Location {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
+
     private String locationName;
     private String description;
 }
